[edr] [LINUXDAR-812] Handle osquery log rotation. add error handling to database purge, check if database purge and log rotation are needed upon start up

+++ edr/modules/pluginimpl/PluginAdapter.cpp
@@ -11,12 +11,10 @@ Copyright 2018-2020 Sophos Limited.  All rights reserved.
 #include "Logger.h"
 #include "TelemetryConsts.h"
 
-#include <Common/FileSystem/IFileSystem.h>
 #include <Common/TelemetryHelperImpl/TelemetryHelper.h>
-#include <boost/property_tree/ini_parser.hpp>
-#include <boost/property_tree/ptree.hpp>
 #include <modules/Proc/ProcUtilities.h>
 #include <Common/FileSystem/IFileSystem.h>
+#include <Common/FileSystem/IFileSystemException.h>
 
 #include <cmath>
 
@@ -60,7 +58,6 @@ public:
     }
 };
 #include "Telemetry.h"
-
 #include <livequery/ResponseDispatcher.h>
 
 namespace Plugin
@@ -74,8 +71,8 @@ namespace Plugin
         m_queueTask(std::move(queueTask)),
         m_baseService(std::move(baseService)),
         m_callback(std::move(callback)),
-        m_queryProcessor { std::move(queryProcessor) },
-        m_responseDispatcher { std::move(responseDispatcher) },
+        m_queryProcessor{std::move(queryProcessor)},
+        m_responseDispatcher{std::move(responseDispatcher)},
         m_timesOsqueryProcessFailedToStart(0)
     {
     }
@@ -83,8 +80,7 @@ namespace Plugin
     void PluginAdapter::mainLoop()
     {
         LOGINFO("Entering the main loop");
-        prepareSystemForPlugin();
-        setUpOsqueryMonitor();
+        cleanAndSetUp();
 
         std::unique_ptr<WaitUpTo> m_delayedRestart;
         while (true)
@@ -92,8 +88,7 @@ namespace Plugin
             Task task;
             if( !m_queueTask->pop(task, 3600))
             {
-                databasePurge();
-                setUpOsqueryMonitor();
+                cleanAndSetUp();
             }
             else
             {
@@ -143,17 +138,80 @@ namespace Plugin
         }
     }
 
+    void PluginAdapter::rotateOsqueryLogs()
+    {
+        auto* ifileSystem = Common::FileSystem::fileSystem();
+        std::string logPath = Plugin::osQueryLogPath();
+        off_t size = ifileSystem->fileSize(logPath);
+
+        if (size > MAX_LOGFILE_SIZE)
+        {
+            stopOsquery();
+            LOGDEBUG("Rotating osquery logs");
+            std::string fileToDelete = Common::FileSystem::join(logPath,".10");
+
+            if (ifileSystem->isFile(fileToDelete))
+            {
+                LOGDEBUG("Log limit reached : Deleting oldest osquery log file");
+                ifileSystem->removeFile(fileToDelete);
+            }
+
+            int iterator = 9;
+            while (iterator > 0)
+            {
+                std::string oldExtension = "."+ std::to_string(9);
+                std::string fileToIncrement = Common::FileSystem::join(logPath,oldExtension);
+
+                if (ifileSystem->isFile(fileToIncrement))
+                {
+                    std::string newExtension = "."+ std::to_string(9+1);
+                    std::string fileDestination = Common::FileSystem::join(logPath,newExtension);
+                    ifileSystem->moveFile(fileToIncrement,fileDestination);
+                }
+
+                iterator -= 1;
+            }
+        }
+    }
+
+    void PluginAdapter::cleanAndSetUp()
+    {
+        databasePurge();
+        rotateOsqueryLogs();
+        setUpOsqueryMonitor();
+    }
+
     void PluginAdapter::databasePurge()
     {
         auto* ifileSystem = Common::FileSystem::fileSystem();
-        std::vector<std::string> paths = ifileSystem->listFiles(Plugin::osQueryDataBasePath());
-        if (paths.size() > 100)
+        try
         {
-            stopOsquery();
-            for (const auto& filepath : paths)
+            LOGSUPPORT("Checking osquery database size");
+            std::string databasePath = Plugin::osQueryDataBasePath();
+            if (ifileSystem->isDirectory(databasePath))
             {
-                ifileSystem->removeFile(filepath);
+                std::vector<std::string> paths = ifileSystem->listFiles(databasePath);
+
+                if (paths.size() > 100) {
+                    LOGINFO("Purging Database");
+                    stopOsquery();
+                    
+                    for (const auto &filepath : paths)
+                    {
+                        ifileSystem->removeFile(filepath);
+                    }
+                }
             }
+            else
+            {
+                LOGSUPPORT("Osquery database does not exist");
+            }
+        }
+        catch (Common::FileSystem::IFileSystemException& e)
+        {
+            std::stringstream errorMessage;
+            errorMessage << "Database cannot be purged due to exception: " << e.what();
+            LOGERROR(errorMessage.str());
         }
     }
 
@@ -225,187 +283,10 @@ namespace Plugin
         // safe to clean up.
     }
 
-    void PluginAdapter::processQuery(const std::string& queryJson, const std::string& correlationId)
+    void PluginAdapter::processQuery(const std::string &queryJson, const std::string &correlationId)
     {
         livequery::processQuery(*m_queryProcessor, *m_responseDispatcher, queryJson, correlationId);
     }
 
-    void PluginAdapter::regenerateOsqueryConfigFile(const std::string& osqueryConfigFilePath)
-    {
-        auto fileSystem = Common::FileSystem::fileSystem();
-        if (fileSystem->isFile(osqueryConfigFilePath))
-        {
-            fileSystem->removeFile(osqueryConfigFilePath);
-        }
-
-        std::stringstream osqueryConfiguration;
-
-        osqueryConfiguration << R"(
-        {
-            "options": {
-                "schedule_splay_percent": 10
-            },
-            "schedule": {
-                "process_events": {
-                    "query": "select count(*) as process_events_count from process_events;",
-                    "interval": 86400
-                },
-                "user_events": {
-                    "query": "select count(*) as user_events_count from user_events;",
-                    "interval": 86400
-                },
-                "selinux_events": {
-                    "query": "select count(*) as selinux_events_count from selinux_events;",
-                    "interval": 86400
-                },
-                "socket_events": {
-                    "query": "select count(*) as socket_events_count from socket_events;",
-                    "interval": 86400
-                }
-            }
-        })";
-
-        fileSystem->writeFile(osqueryConfigFilePath, osqueryConfiguration.str());
-    }
-
-    void PluginAdapter::regenerateOSQueryFlagsFile(const std::string& osqueryFlagsFilePath)
-    {
-        auto fileSystem = Common::FileSystem::fileSystem();
-
-        if (fileSystem->isFile(osqueryFlagsFilePath))
-        {
-            fileSystem->removeFile(osqueryFlagsFilePath);
-        }
-
-        std::vector<std::string> flags { "--host_identifier=uuid",
-                                         "--log_result_events=true",
-                                         "--utc",
-                                         "--disable_extensions=false",
-                                         "--logger_stderr=false",
-                                         "--logger_mode=420",
-                                         "--logger_min_stderr=1",
-                                         "--logger_min_status=1",
-                                         "--disable_watchdog=false",
-                                         "--watchdog_level=0",
-                                         "--watchdog_memory_limit=250",
-                                         "--watchdog_utilization_limit=30",
-                                         "--watchdog_delay=60",
-                                         "--enable_extensions_watchdog=false",
-                                         "--disable_extensions=false",
-                                         "--disable_audit=false",
-                                         "--audit_persist=true",
-                                         "--enable_syslog=true",
-                                         "--audit_allow_config=true",
-                                         "--audit_allow_process_events=true",
-                                         "--audit_allow_fim_events=false",
-                                         "--audit_allow_selinux_events=true",
-                                         "--audit_allow_sockets=true",
-                                         "--audit_allow_user_events=true",
-                                         "--syslog_events_expiry=604800",
-                                         "--events_expiry=604800",
-                                         "--force=true",
-                                         "--disable_enrollment=true",
-                                         "--enable_killswitch=false",
-                                         "--events_max=20000" };
-
-        flags.push_back("--syslog_pipe_path=" + Plugin::syslogPipe());
-        flags.push_back("--pidfile=" + Plugin::osqueryPidFile());
-        flags.push_back("--database_path=" + Plugin::osQueryDataBasePath());
-        flags.push_back("--extensions_socket=" + Plugin::osquerySocket());
-        flags.push_back("--logger_path=" + Plugin::osQueryLogPath());
-
-        std::ostringstream flagsAsString;
-        std::copy(flags.begin(), flags.end(), std::ostream_iterator<std::string>(flagsAsString, "\n"));
-
-        fileSystem->writeFile(osqueryFlagsFilePath, flagsAsString.str());
-    }
-
-    bool PluginAdapter::checkIfServiceActive(const std::string& serviceName)
-    {
-        auto process = Common::Process::createProcess();
-        process->exec(Plugin::systemctlPath(), { "is-active", serviceName });
-
-        return (process->exitCode() == 0);
-    }
-
-    void PluginAdapter::stopSystemService(const std::string& serviceName)
-    {
-        auto process = Common::Process::createProcess();
-        process->exec(Plugin::systemctlPath(), { "stop", serviceName });
-
-        if (process->exitCode() == 4)
-        {
-            // handle error: unit auditd.service may be requested by dependency only
-            // try to stop the service again using service command
-            process->exec(Plugin::servicePath(), { serviceName, "stop" });
-        }
-
-        if (process->exitCode() == 0)
-        {
-            LOGINFO("Successfully stopped service: " << serviceName);
-            process->exec(Plugin::systemctlPath(), { "disable", serviceName });
-            if (process->exitCode() == 0)
-            {
-                LOGINFO("Successfully disabled service: " << serviceName);
-            }
-            else
-            {
-                LOGWARN("Failed to disable service: " << serviceName);
-            }
-        }
-        else
-        {
-            LOGWARN("Failed to stop service: " << serviceName);
-        }
-    }
-
-    void PluginAdapter::prepareSystemForPlugin()
-    {
-        auto fileSystem = Common::FileSystem::fileSystem();
-        bool disableAuditD = true;
-
-        if (fileSystem->isFile(Plugin::edrConfigFilePath()))
-        {
-            try
-            {
-                boost::property_tree::ptree ptree;
-                boost::property_tree::read_ini(Plugin::edrConfigFilePath(), ptree);
-                disableAuditD = (ptree.get<std::string>("disable_auditd") == "1");
-            }
-            catch (boost::property_tree::ptree_error& ex)
-            {
-                LOGWARN("Failed to read disable_auditd configuration from config file, using default value");
-            }
-        }
-        else
-        {
-            LOGWARN(
-                "Could not find EDR Plugin config file: " << Plugin::edrConfigFilePath()
-                                                          << ", using disable_auditd default value");
-        }
-
-        std::string serviceName("auditd");
-        if (disableAuditD)
-        {
-            if (checkIfServiceActive(serviceName))
-            {
-                stopSystemService(serviceName);
-            }
-            else
-            {
-                LOGINFO("AuditD not found on system or is not active.");
-            }
-        }
-        else
-        {
-            if (checkIfServiceActive(serviceName))
-            {
-                LOGWARN("EDR configuration set to not disable AuditD, it will not be possible to obtain event data.");
-            }
-        }
-
-        regenerateOSQueryFlagsFile(Plugin::osqueryFlagsFilePath());
-        regenerateOsqueryConfigFile(Plugin::osqueryConfigFilePath());
-    }
 
 } // namespace Plugin
